coravel implementation

+++ Personnel.Model/ViewModel/RequestModel/LeaveServiceApproveModel.cs
@@ -6,7 +6,7 @@ namespace Personnel.Model.ViewModel.RequestModel
 {
     public class LeaveServiceApproveModel
     {
-        public decimal BonusAmount { get; set; }
+        public double BonusAmount { get; set; }
         public string ApprovedBy { get; set; }
         public DateTime ApprovedDate { get; set; }
     }
